perf: lift the frame rate on Mali-450 with tunable quality reductions
Paired with the aurora-side FIFO work in the submodule. The port side adds the
measurement and the knobs that buy back frames where the cost is concentrated:

- Frame anatomy behind PARTYBOARD_PERF_SPLIT, splitting events, logic, Hu3D, render
  submission, end-of-frame and limiter. Complements aurora's own [fps] stall split,
  so a slow frame can be attributed to a side before anything is changed.
- Shadow updates can run every Nth frame (PARTYBOARD_SHADOW_DIVISOR, and a separate
  divisor for the heavy minigames), since shadow regeneration cost scales with a
  scene the board rarely changes between frames.
- Mr. Blizzard reflections can be dropped and its particle count divided, the two
  costs that made that minigame the worst case.
- Seed the framebuffer scale before aurora creates its render targets. Applying it
  afterwards forced a live target resize that contends for Mali's EGL context.

Every reduction defaults to full quality and is opt-in, so the shipped build is
unchanged unless a knob is set.

+++ include/port/settings.h
@@ -86,7 +86,7 @@ struct UserSettings {
         ConfigVar<bool> enableControllerToasts;
 
         // Graphics
-        ConfigVar<int> internalResolutionScale;
+        ConfigVar<float> internalResolutionScale;
         ConfigVar<int> shadowResolutionMultiplier;
 
         // Audio
